backend/pyproject.toml
[project]
name = "web-shop-django"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-debug-toolbar (>=5.1.0,<6.0.0)",
    "django-bootstrap5 (>=25.1,<26.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-minio-storage (>=0.5.7,<0.6.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
